Moved components to subfolder within envs and updated all files
