+++ admiral/envs/components/movement.py
@@ -2,7 +2,7 @@
 import numpy as np
 
 from admiral.envs import Agent
-from admiral.component_envs.position import GridPositionAgent
+from admiral.envs.components.position import GridPositionAgent
 
 class GridMovementAgent(Agent):
     """
