+++ examples/component_examples/predator_prey_example.py
@@ -1,15 +1,15 @@
 
 from matplotlib import pyplot as plt
 
-from admiral.component_envs.observer import ObservingAgent
-from admiral.component_envs.team import TeamAgent
-from admiral.component_envs.position import GridPositionTeamsComponent, GridPositionAgent
-from admiral.component_envs.movement import GridMovementComponent, GridMovementAgent
-from admiral.component_envs.attacking import GridAttackingTeamComponent, GridAttackingAgent
-from admiral.component_envs.death_life import DyingComponent, LifeAgent
-from admiral.component_envs.resources import GridResourceComponent, GridResourceHarvestingAgent
-from admiral.component_envs.rewarder import RewarderComponent
-from admiral.component_envs.done_component import TeamDeadDoneComponent
+from admiral.envs.components.observer import ObservingAgent
+from admiral.envs.components.team import TeamAgent
+from admiral.envs.components.position import GridPositionTeamsComponent, GridPositionAgent
+from admiral.envs.components.movement import GridMovementComponent, GridMovementAgent
+from admiral.envs.components.attacking import GridAttackingTeamComponent, GridAttackingAgent
+from admiral.envs.components.death_life import DyingComponent, LifeAgent
+from admiral.envs.components.resources import GridResourceComponent, GridResourceHarvestingAgent
+from admiral.envs.components.rewarder import RewarderComponent
+from admiral.envs.components.done_component import TeamDeadDoneComponent
 from admiral.envs import AgentBasedSimulation
 
 class PreyAgent(GridPositionAgent, ObservingAgent, TeamAgent, GridMovementAgent, LifeAgent, GridResourceHarvestingAgent):
